+++ uart/ns16550a/src/driver/server.rs
@@ -127,7 +127,7 @@ impl<'a> SystemService for UartService<'a> {
                     let slot = SHM_SLOT;
                     let vaddr = u.get_mr(0);
                     let size = u.get_mr(1);
-                    let paddr = u.get_mr(2) as u64;
+                    let paddr = u.get_mr(2) as usize;
                     CSPACE_CAP.move_cap(recv_slot, slot)?;
                     let frame = Frame::from(slot);
                     s.setup_shm(frame, vaddr, paddr, size)?;
